fix: use float precision for time in LivingCanvas shader
half precision for _Time.y breaks down after ~17 minutes as the value
exceeds half's representable range, causing particle positions to step
in discrete jumps instead of moving smoothly.

+++ Assets/Shaders/LivingCanvas.shader
@@ -119,7 +119,7 @@ Shader "Garden/LivingCanvas"
             half4 frag(Varyings input) : SV_Target
             {
                 float2 uv = input.uv;
-                half t = _Time.y;
+                float t = _Time.y;
 
                 // --- 1. Vertical gradient ---
                 half3 col = lerp(_BottomColor.rgb, _TopColor.rgb, uv.y);
